+++ CHANGES.md
@@ -6,7 +6,7 @@
 * Add an alway-on mode. (Suggested by [Dan Sayers](https://plus.google.com/108024724934415857156) and [Alexander Sergéev](https://github.com/sergeevabc)).
 * Add a magnifier which helps user to spot-check the page while moving mouse on the preview bar. (Experimental) (Suggested by [Dan Sayers](https://plus.google.com/108024724934415857156)).
 * Fix precision of click-to-jump. Now, you can check the magnifier and click to jump exactly where you found.
-
+* Apply anti-aliasing on snapshot. (Suggested by [modi X](https://plus.google.com/115562168718511548938))
 
 ## Version 1.0